Lot of stuff on todo

+++ Pancakes/GenericBlockCell.m
@@ -355,7 +355,7 @@
     [closeButton setImage:[UIImage imageNamed:@"block-close"] forState:UIControlStateNormal];
     closeButton.backgroundColor = [Utils colorWithHexString:self.articleViewController.displayedArticle.color];
     closeButton.layer.cornerRadius = 20;
-    closeButton.tag = blockId;
+    closeButton.tag = blockId.intValue;
     [closeButton addTarget:self action:@selector(closeButtonTapped:) forControlEvents:UIControlEventTouchUpInside];
     [self addSubview:closeButton];
